refactor(client): :recycle: Rename LocationDropdown to LocationMenu
