+++ client/src/components/Pages/homepage/HomeSearchHeader/HomeSearchHeader.desktop.tsx
@@ -1,4 +1,4 @@
-import LocationDropdown from "components/Pages/recherche/LocationDropdown";
+import LocationMenu from "components/Pages/recherche/LocationMenu";
 import ThemeDropdown from "components/Pages/recherche/ThemeDropdown";
 import EVAIcon from "components/UI/EVAIcon/EVAIcon";
 import { cls } from "lib/classname";
@@ -79,7 +79,7 @@ const HomeSearchHeaderDesktop = (props: Props) => {
           values={query.departments}
           resetFilter={resetDepartment}
         >
-          <LocationDropdown />
+          <LocationMenu />
         </SearchDropdown>
 
         <SearchDropdown
