--- client/src/components/Pages/recherche/LocationDropdown/index.ts
@@ -1,3 +1,0 @@
-import LocationDropdown from "./LocationDropdown";
-
-export default LocationDropdown;
