+++ client/src/components/Pages/recherche/LocationMenu/functions.ts
