+++ client/src/components/Pages/recherche/LocationMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LocationMenu";
