+++ client/src/components/Pages/recherche/SearchHeader/Filters.tsx
@@ -6,7 +6,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { Container } from "reactstrap";
 import { addToQueryActionCreator } from "services/SearchResults/searchResults.actions";
 import { searchQuerySelector, themesDisplayedValueSelector } from "services/SearchResults/searchResults.selector";
-import LocationDropdown from "../LocationDropdown";
+import LocationMenu from "../LocationMenu";
 import ThemeDropdown from "../ThemeDropdown";
 import Filter from "./Filter";
 import styles from "./Filters.module.scss";
@@ -65,7 +65,7 @@ const Filters = (props: Props) => {
           dropdownMenu={{
             value: query.departments,
             reset: resetDepartment,
-            menu: <LocationDropdown />,
+            menu: <LocationMenu />,
           }}
           gaType="department"
         />
